refactor: move fetchToCurl to generateCodeSampleCurl
